# wsl-docker-service/internal/httpserver/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "docklite-wsl",
    "description": "HTTP API cho DockLite (Docker/Compose trong WSL). Khi biến môi trường DOCKLITE_API_TOKEN khác rỗng, mọi yêu cầu cần header Authorization: Bearer <token> hoặc X-DockLite-Token. Header tùy chọn X-Request-ID (chuẩn hóa trên server): client gửi ID tương quan; response trả cùng giá trị; log slog dùng req_id. WebSocket: giới hạn số kết nối đồng thời bằng DOCKLITE_WS_MAX_CONNECTIONS (mặc định 64). Tài liệu này mô tả route và phương thức; chi tiết body/query xem mã nguồn và README.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Tương đối ServiceBaseUrl (ví dụ http://127.0.0.1:17890/)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Khớp DOCKLITE_API_TOKEN khi bật xác thực"
      },
      "dockLiteToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-DockLite-Token",
        "description": "Thay thế tùy chọn cho Bearer"
      }
    },
    "schemas": {
      "ComposeIdBody": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "string",
            "description": "ID project Compose đã lưu (UUID)"
          },
          "profiles": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Tùy chọn: lặp lại `docker compose --profile <tên>` (tối đa 32 profile, mỗi tên tối đa 128 ký tự)"
          }
        }
      },
...
      "AuthRotateRequest": {
        "type": "object",
        "required": ["current_token"],
        "properties": {
          "current_token": {
            "type": "string",
            "description": "Mật khẩu Bearer hiện tại (phải khớp với token trên server)"
          }
        }
      },
...
    },
    "responses": {
      "OkEnvelope": {
        "description": "Thành công (envelope JSON apiresponse)",
        "content": {
          "application/json": {
            "schema": { "type": "object" }
          }
        }
      },
      "Unauthorized": {
        "description": "Thiếu hoặc sai token khi DOCKLITE_API_TOKEN được đặt"
      }
    }
  },
  "paths": {
...
    "/api/auth/rotate": {
      "post": {
        "tags": ["system"],
        "summary": "Xoay mật khẩu API (Bearer)",
        "description": "Chỉ đăng ký khi DOCKLITE_API_TOKEN khác rỗng lúc khởi động. Giới hạn khoảng 5 yêu cầu/phút theo IP.",
        "security": [{ "bearerAuth": [] }, { "dockLiteToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthRotateRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/OkEnvelope" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "Xác thực API chưa bật trên service" },
          "429": { "description": "Quá nhiều yêu cầu xoay token" }
        }
      }
    },
...
    "/api/compose/up": {
      "post": {
        "tags": ["compose"],
        "summary": "compose up",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ComposeIdBody" }
            }
          }
        },
        "responses": { "200": { "$ref": "#/components/responses/OkEnvelope" } }
      }
    },
    "/api/compose/down": {
      "post": {
        "tags": ["compose"],
        "summary": "compose down",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ComposeIdBody" }
            }
          }
        },
        "responses": { "200": { "$ref": "#/components/responses/OkEnvelope" } }
      }
    },
...
  }
}